requirements.txt
alabaster==0.7.16
asgiref==3.8.1
Babel==2.15.0
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
Django==5.0.6
django-cors-headers==4.3.1
django-storages==1.14.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
docutils==0.21.2
idna==3.7
imagesize==1.4.1
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.0
pillow==10.2.0
psycopg==3.1.18
Pygments==2.18.0
PyJWT==2.8.0
requests==2.31.0
snowballstemmer==2.2.0
Sphinx==7.3.7
sphinx-markdown-builder==0.6.6
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
sqlparse==0.4.4
tabulate==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
whitenoise==6.6.0
